requirements.txt
alembic==1.5.2
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.11.0
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
login==0.0.6
Mako==1.1.4
MarkupSafe==1.1.1
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
